Refactor namespaces and update dependencies
Updated `CustomCopilot.csproj` to upgrade package versions for `Azure.Maps.Search`, `Microsoft.SemanticKernel`, and `Microsoft.SemanticKernel.Plugins.Core`. Renamed namespaces and classes in `FlightTracker.cs`, `PlaceSuggestions.cs`, and `Weather.cs` to remove the "Plugin" suffix. Updated `PlaceSuggestions.cs` to use `GetGeocodingAsync` instead of `SearchAddressAsync`. Modified `Program.cs` to reflect new class names, updated Azure OpenAI chat completion service configuration, removed unnecessary `using` directives, adjusted chat history initialization, and updated `PromptExecutionSettings` initialization to use `FunctionChoiceBehavior.Auto()`.
