+++ samples/dotnet/CustomCopilot/CustomCopilot.csproj
@@ -8,9 +8,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Azure.Maps.Search" Version="1.0.0-beta.4" />
-    <PackageReference Include="Microsoft.SemanticKernel" Version="1.3.1" />
-    <PackageReference Include="Microsoft.SemanticKernel.Plugins.Core" Version="1.3.1-alpha" />
+    <PackageReference Include="Azure.Maps.Search" Version="2.0.0-beta.4" />
+    <PackageReference Include="Microsoft.SemanticKernel" Version="1.26.0" />
+    <PackageReference Include="Microsoft.SemanticKernel.Plugins.Core" Version="1.26.0-alpha" />
   </ItemGroup>
 
 </Project>
